Move preview page data getter test to summaries module suite
