+++ budgetportal/tests/test_department.py
@@ -9,14 +9,13 @@ from budgetportal.models import (
     Government,
     Department,
 )
-from budgetportal import models, datasets
+from budgetportal import models
 from django.test import TestCase
 from mock import Mock
 import json
-import requests
-from budgetportal.openspending import BabbageFiscalDataset
 
 # Hacky make sure we don't call out to openspending.
+import requests
 requests.get = Mock
 requests.Session = Mock
 
@@ -27,9 +26,6 @@ with open('budgetportal/tests/test_data/budget_and_actual.json', 'r') as f:
 
 with open('budgetportal/tests/test_data/test_treemap_expenditure_national.json', 'r') as f:
     TREEMAP_MOCK_DATA = json.load(f)
-
-with open('budgetportal/tests/test_data/test_national_department_preview.json', 'r') as f:
-    NATIONAL_DEPARTMENT_PREVIEW_MOCK_DATA = json.load(f)
 
 
 class AdjustedBudgetMissingTestCase(TestCase):
@@ -361,75 +357,3 @@ class NationalTreemapExpenditureByDepartmentTestCase(TestCase):
         self.assertIn('percentage_of_total', expenditure_keys)
         self.assertIn('url', expenditure_keys)
         self.assertIn('province', expenditure_keys)
-
-
-class NationalDepartmentPreviewTestCase(TestCase):
-    """ Unit tests for the national department preview department function. """
-
-    def setUp(self):
-        self.mock_data = NATIONAL_DEPARTMENT_PREVIEW_MOCK_DATA
-        year = FinancialYear.objects.create(slug="2019-20")
-        sphere = Sphere.objects.create(financial_year=year, name="national")
-        government = Government.objects.create(sphere=sphere, name="South Africa")
-        self.department = Department(
-            government=government,
-            name="Fake",
-            vote_number=1,
-            is_vote_primary=True,
-            intro="",
-        )
-        mock_dataset = Mock()
-        self.mock_openspending_api = Mock()
-        self.mock_openspending_api.get_adjustment_kind_ref = Mock(return_value='adjustment_kind_ref')
-        self.mock_openspending_api.get_phase_ref = Mock(return_value='budget_phase.budget_phase')
-        self.mock_openspending_api.get_programme_name_ref = Mock(return_value='programme_number.programme')
-        self.mock_openspending_api.get_department_name_ref = Mock(return_value='vote_number.department')
-        self.mock_openspending_api.get_geo_ref = Mock(return_value='geo_source.government')
-        self.mock_openspending_api.get_function_ref = Mock(return_value='function_group_1.function_group_1')
-        self.mock_openspending_api.get_financial_year_ref = Mock(return_value="financial_year.financial_year")
-        self.mock_openspending_api.aggregate = Mock(return_value={'cells': self.mock_data['programmes']})
-        self.mock_openspending_api.filter_by_ref_exclusion = Mock(return_value=self.mock_data['programmes'])
-        self.mock_openspending_api.aggregate_by_refs = BabbageFiscalDataset.aggregate_by_refs
-        self.mock_openspending_api.aggregate_url = Mock
-        mock_dataset.get_openspending_api = Mock(return_value=self.mock_openspending_api)
-        self.department.get_expenditure_time_series_dataset = Mock(return_value=mock_dataset)
-
-        vote_number = 1
-        for mock_object in self.mock_data['departments']:
-            Department.objects.create(
-                government=government,
-                is_vote_primary=True,
-                name=mock_object['vote_number.department'],
-                vote_number=vote_number
-            )
-            vote_number += 1
-
-    @mock.patch(
-        'budgetportal.models.Department.get_all_budget_totals_by_year_and_phase',
-        return_value=mock.MagicMock()
-    )
-    def test_no_cells_null_response(self, total_budgets_mock):
-        self.mock_openspending_api.aggregate_by_refs = Mock(return_value=[])
-        result = self.department.get_national_expenditure_treemap(
-            financial_year_id='2018-19',
-            budget_phase='original'
-        )
-        self.assertEqual(result, None)
-
-    @mock.patch('budgetportal.models.Department.get_all_budget_totals_by_year_and_phase', return_value=mock.MagicMock())
-    def test_complete_data(self, total_budgets_mock):
-        result = self.department.get_preview_page(financial_year_id='2019-20', phase_slug='original',
-                                                  government_slug='south-africa', sphere_slug='national')
-        data = result['data']
-        self.assertEqual(len(data), 1)
-        data_keys = data.keys()
-        self.assertIn('items', data_keys)
-        expenditure_keys = data['items'][0].keys()
-        self.assertIn('title', expenditure_keys)
-        self.assertIn('description', expenditure_keys)
-        self.assertIn('percentage_of_budget', expenditure_keys)
-        self.assertIn('programmes', expenditure_keys)
-        self.assertIn('slug', expenditure_keys)
-        self.assertIn('focus_areas', expenditure_keys)
-
-        self.assertEqual(data['items'][0]['focus_areas'][0]['slug'], 'economic-development')
